imporved bash script, moved env type to config dir
