--- client/.env
@@ -1,5 +1,0 @@
-REACT_APP_SERVER_HOSTNAME =  'localhost';
-SERVER_PORT =  3500;
-REACT_APP_URL_LIB = https://ingenioer.au.dk/
-REACT_APP_URL_DT = https://jupyterlite.github.io/demo/repl/index.html
-REACT_APP_URL_WORKBENCH = https://ece.au.dk/
