+++ client/.eslintignore
@@ -2,3 +2,4 @@ build/
 node_modules/
 script/
 api/
+config/
