+++ client/config/prod.js
@@ -1,4 +1,6 @@
-window.env  = {
-  "REACT_APP_ENVIRONMENT" : "prod",
-  "REACT_APP_JUPYTER_URL" : "https://jupyterlite.github.io/demo/repl/index.html?kernel=javascript&toolbar="
-}
+window.env = {
+  REACT_APP_ENVIRONMENT: 'prod',
+  REACT_APP_URL_LIB: 'https://ingenioer.au.dk/',
+  REACT_APP_URL_DT: 'https://au.dk/',
+  REACT_APP_URL_WORKBENCH: 'https://ece.au.dk/',
+};
