+++ client/script/config.bash
@@ -2,11 +2,28 @@
 # https://stackoverflow.com/questions/51653931/react-configuration-file-for-post-deployment-settings
 # https://dev.to/akdevcraft/react-runtime-variables-49dc
 
-mode=${REACT_APP_ENV-dev}
+mode=$1
+if [ -z "$mode" ]; then
+    echo "No mode specified - use 'dev' or 'prod'"
+    exit 1
+fi
+
+set_env() {
+    echo "Setting env to $1"
+    cp config/$1.js public/env.js # Configure dev environment in public for next build
+    if [ -d build ]; then
+        cp public/env.js build/env.js # Hot swap dev to build
+    fi
+}
 
 case "$mode" in 
-    dev) cp config/dev.js build/static/env.js
-         ;;
-    prod) cp config/prod.js build/static/env.js
+    dev)
+        set_env dev
+        ;;
+    prod)
+        set_env prod
+        ;;
+    *) echo "Invalid mode $mode - use 'dev' or 'prod'"
+        exit 1
         ;;
 esac
